iox-#563 move tests in posh integrationtests into anonymous namespace
