+++ iceoryx_posh/test/integrationtests/test_interface_port_stack_blowup.cpp
@@ -22,6 +22,8 @@
 
 #include "test.hpp"
 
+namespace
+{
 using namespace ::testing;
 using ::testing::Return;
 
@@ -42,3 +44,5 @@ TEST_F(InterfacePortRequestStackBlowup_test, RouDiMustContinue)
     // we don't care if there are capro messages or not, we just want to have a check that there was no segfault
     EXPECT_THAT(sut.getCaProMessage(caproMessage), AnyOf(true, false));
 }
+
+} // namespace
