+++ iceoryx_posh/test/integrationtests/test_roudi_findservice.cpp
@@ -1,4 +1,5 @@
 // Copyright (c) 2019, 2021 by Robert Bosch GmbH. All rights reserved.
+// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,10 +16,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "iceoryx_posh/runtime/posh_runtime.hpp"
+#include "iceoryx_posh/testing/roudi_gtest.hpp"
 #include "iceoryx_utils/cxx/convert.hpp"
-#include "test_roudi_service_discovery.hpp"
 
-class RoudiFindService_test : public RouDiServiceDiscoveryTest
+namespace
+{
+using iox::capro::IdString_t;
+using iox::runtime::InstanceContainer;
+
+class RoudiFindService_test : public RouDi_GTest
 {
   public:
     void SetUp()
@@ -27,6 +33,15 @@ class RoudiFindService_test : public RouDiServiceDiscoveryTest
 
     void TearDown()
     {
+    }
+
+    void InitContainer(InstanceContainer& dest, std::vector<std::string> src)
+    {
+        dest.clear();
+        for (size_t i = 0; i < src.size(); i++)
+        {
+            dest.push_back(IdString_t(iox::cxx::TruncateToCapacity, src[i]));
+        }
     }
 
     iox::runtime::PoshRuntime* senderRuntime{&iox::runtime::PoshRuntime::initRuntime("sender")};
@@ -397,3 +412,5 @@ TEST_F(RoudiFindService_test, findServiceInstanceContainerOverflowError)
 
     ASSERT_THAT(instanceContainer.has_error(), Eq(true));
 }
+
+} // namespace
